requirements.txt
streamlit
gspread
oauth2client
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2